+++ packages/scenes/src/components/layout/grid/types.ts
@@ -1,0 +1,15 @@
+import { SceneObject, SceneObjectStatePlain } from '../../../core/types';
+
+export interface SceneGridItemPlacement {
+  x?: number;
+  y?: number;
+  width?: number;
+  height?: number;
+}
+
+export interface SceneGridItemStateLike extends SceneGridItemPlacement, SceneObjectStatePlain {
+  isResizable: boolean;
+  isDraggable: boolean;
+}
+
+export type SceneGridItemLike = SceneObject<SceneGridItemStateLike>;