+++ packages/scenes/src/core/SceneTimeRange.test.tsx
@@ -12,12 +12,12 @@ describe('SceneTimeRange', () => {
     await new Promise((r) => setTimeout(r, 2));
     timeRange.onRefresh();
     const diff = timeRange.state.value.from.valueOf() - startTime;
-    expect(diff).toBeGreaterThan(1);
+    expect(diff).toBeGreaterThan(0);
   });
 
   it('toUrlValues with relative range', () => {
     const timeRange = new SceneTimeRange({ from: 'now-1h', to: 'now' });
-    expect(timeRange.urlSync?.getUrlState(timeRange.state)).toEqual({
+    expect(timeRange.urlSync?.getUrlState()).toEqual({
       from: 'now-1h',
       to: 'now',
     });