+++ packages/scenes/src/core/types.ts
@@ -25,6 +25,10 @@ export interface SceneLayoutChildState extends SceneObjectStatePlain {
   placement?: SceneLayoutChildOptions;
 }
 
+export interface SceneLayoutItemState extends SceneObjectStatePlain {
+  body: SceneObject | undefined;
+}
+
 export type SceneObjectState = SceneObjectStatePlain | SceneLayoutState | SceneLayoutChildState;
 
 export interface SceneLayoutChildOptions {
@@ -70,7 +74,7 @@ export interface SceneObject<TState extends SceneObjectState = SceneObjectState>
   readonly variableDependency?: SceneVariableDependencyConfigLike;
 
   /** This abstraction declares URL sync dependencies of a scene object. **/
-  readonly urlSync?: SceneObjectUrlSyncHandler<TState>;
+  readonly urlSync?: SceneObjectUrlSyncHandler;
 
   /** Subscribe to state changes */
   subscribeToState(handler: SceneStateChangedHandler<TState>): Unsubscribable;
@@ -123,7 +127,7 @@ export type SceneDeactivationHandler = () => void;
 export type SceneLayoutChild = SceneObject<SceneLayoutChildState | SceneLayoutState>;
 
 export interface SceneLayoutState extends SceneLayoutChildState {
-  children: SceneLayoutChild[];
+  children: SceneObject[];
 }
 
 export interface SceneLayout<T extends SceneLayoutState = SceneLayoutState> extends SceneObject<T> {
@@ -153,14 +157,14 @@ export function isSceneObject(obj: any): obj is SceneObject {
   return obj.useState !== undefined;
 }
 
-export interface SceneObjectWithUrlSync<TState> extends SceneObject {
-  getUrlState(state: TState): SceneObjectUrlValues;
+export interface SceneObjectWithUrlSync extends SceneObject {
+  getUrlState(): SceneObjectUrlValues;
   updateFromUrl(values: SceneObjectUrlValues): void;
 }
 
-export interface SceneObjectUrlSyncHandler<TState> {
+export interface SceneObjectUrlSyncHandler {
   getKeys(): string[];
-  getUrlState(state: TState): SceneObjectUrlValues;
+  getUrlState(): SceneObjectUrlValues;
   updateFromUrl(values: SceneObjectUrlValues): void;
 }