+++ packages/scenes/CHANGELOG.md
@@ -1,3 +1,94 @@
+# v0.2.0 (Wed Mar 29 2023)
+
+### Release Notes
+
+#### Layout: Create atomic, layout specific objects ([#97](https://github.com/grafana/scenes/pull/97))
+
+The interface of `SceneFlexLayout` and `SceneGridLayout` has changed. These scene objects now accept only dedicated layout item objects as children:
+- `SceneFlexItem` for `SceneFlexLayout`
+- `SceneGridItem` and `SceneGridRow` for `SceneGridLayout`
+
+`placement` property has been replaced by those layout-specific objects. 
+
+Example
+```tsx
+// BEFORE
+const layout = new SceneFlexLayout({
+  direction: 'column',
+  children: [
+    new VizPanel({
+      placement: {
+        width: '50%',
+        height: '400',
+     },
+     ...
+    })
+  ],
+  ...
+})
+
+
+// AFTER
+const layout = new SceneFlexLayout({
+  direction: 'column',
+  children: [ 
+    new SceneFlexItem({ 
+      width: '50%',
+      height: '400',
+      body: new VizPanel({ ... }),
+    }),
+  ],
+  ...
+})
+
+```
+
+---
+
+#### 🚀 Enhancement
+
+- Layout: Create atomic, layout specific objects [#97](https://github.com/grafana/scenes/pull/97) ([@dprokop](https://github.com/dprokop) [@torkelo](https://github.com/torkelo))
+- Interpolation: FormatRegistryID is now replaced by VariableFormatID from schema package [#112](https://github.com/grafana/scenes/pull/112) ([@ryantxu](https://github.com/ryantxu) [@torkelo](https://github.com/torkelo))
+
+#### Authors: 3
+
+- Dominik Prokop ([@dprokop](https://github.com/dprokop))
+- Ryan McKinley ([@ryantxu](https://github.com/ryantxu))
+- Torkel Ödegaard ([@torkelo](https://github.com/torkelo))
+
+---
+
+# v0.1.0 (Mon Mar 27 2023)
+
+### Release Notes
+
+#### UrlSync: Simplify url sync interface ([#100](https://github.com/grafana/scenes/pull/100))
+
+The  SceneObjectUrlSyncHandler interface has changed. The function `getUrlState` no longer takes state as parameter. The implementation needs to use the current scene object state instead.
+
+---
+
+#### 🚀 Enhancement
+
+- UrlSync: Simplify url sync interface [#100](https://github.com/grafana/scenes/pull/100) ([@torkelo](https://github.com/torkelo))
+
+#### 🐛 Bug Fix
+
+- Clean up changelog [#108](https://github.com/grafana/scenes/pull/108) ([@torkelo](https://github.com/torkelo))
+
+#### Authors: 1
+
+- Torkel Ödegaard ([@torkelo](https://github.com/torkelo))
+
+---
+
+# v0.0.32 (Mon Mar 27 2023)
+
+- Scene: Support for new types of "macro" variables starting with \_\_all_variables [#98](https://github.com/grafana/scenes/pull/98) ([@domasx2](https://github.com/domasx2) [@torkelo](https://github.com/torkelo))
+- UrlSyncManager: Improvements and fixes [#96](https://github.com/grafana/scenes/pull/96) ([@torkelo](https://github.com/torkelo))
+
+* UrlSync: SceneObject that implement url sync \_urlSync property will now see a change to how updateFromUrl is called. It is now called with null values when url query parameters are removed. Before the UrlSyncManager would remember the initial state and pass that to updateFromUrl, but now if you want to preserve your current state or set to some initial state you have to handle that logic inside updateFromUrl.
+
 # v0.0.28 (Tue Mar 21 2023)
 
 - Removal of isEditing from SceneComponentProps (also $editor from SceneObjectState, and sceneGraph.getSceneEditor)